+++ SeriesYPeliculas/ModelData/Response/UpcomingResponse/UpcomingResults.swift
@@ -14,7 +14,7 @@ struct UpcomingResults : Codable {
     var release_date : String?
     var title : String?
     var video : Bool?
-    var vote_average : Int?
+    var vote_average : Double?
     var vote_count : Int?
 
 }
